+++ src/models/acl.go
@@ -1,0 +1,9 @@
+package models
+
+// AclRule represents an access control rule
+type AclRule struct {
+	User   string `json:"user"`
+	Item   string `json:"item"`
+	Action string `json:"action"`
+	Type   string `json:"type"`
+}